Add CampaignMapper and unit tests for various controllers
- Implement CampaignMapper for converting between MarketingCampaign entities and CampaignDto.
- Add unit tests for ContractsController covering CRUD operations, approval, and termination workflows.
- Add unit tests for InvoicesController covering CRUD operations, status management, and voiding.
- Add unit tests for LeadsController covering CRUD operations, conversion, and pagination.
- Add unit tests for PaymentsController covering CRUD operations, status management, and completion.
- Add unit tests for QuotesController covering CRUD operations, acceptance, and rejection.

+++ CRM.Backend/tests/CRM.Tests/Integration/Services/ActivityServiceTests.cs
@@ -77,11 +77,11 @@ public class ActivityServiceTests
         var service = new ActivityService(context, _mockLogger.Object);
 
         // Act
-        var result = await service.GetActivitiesAsync(activityType: ActivityType.ChatMessage);
+        var result = await service.GetActivitiesAsync(activityType: (int)ActivityType.ChatMessage);
 
         // Assert
         Assert.Equal(2, result.Count());
-        Assert.All(result, a => Assert.Equal(ActivityType.ChatMessage, a.ActivityType));
+        Assert.All(result, a => Assert.Equal((int)ActivityType.ChatMessage, a.ActivityType));
     }
 
     [Fact]
@@ -282,7 +282,7 @@ public class ActivityServiceTests
 
         // Assert
         Assert.Equal(2, result.Count());
-        Assert.Contains(result, a => a.ActivityType == ActivityType.ChatMessage);
+        Assert.Contains(result, a => a.ActivityType == (int)ActivityType.ChatMessage);
     }
 
     [Fact]
